+++ packages/nameguard-sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @namehash/nameguard
 
+## 0.8.0
+
+### Minor Changes
+
+- 1891ce7: release
+
+### Patch Changes
+
+- Updated dependencies [1891ce7]
+- Updated dependencies [c748998]
+  - @namehash/ens-utils@1.19.0
+
+## 0.7.0
+
+### Minor Changes
+
+- 47de3ad: - Rename `ImpersonationStatus` to `ImpersonationEstimate` to better manage expectations.
+  - Rename `endpoint` param to `nameguardEndpoint` when creating a NameGuard Client for more self-documenting code.
+  - Refined unit tests.
+
 ## 0.6.0
 
 ### Minor Changes